relocated and renamed the e2e test files
