+++ src/app/tests/pages/RegisterPageObject.js
